Cleanup of working RRTM LW code with Python i/o.
